ecwrite_py - If access to the EC cannot be obtained then the app will now shutdown.
frontend.py
- Bug fix, cast tick internal value to int.

main.py
- If an empty string is returned when querying the voltage, skip updating. Prevents crashes when msr-tools is not installed.

- Updated README to add more documentation on installation.

+++ frontend.py
@@ -115,7 +115,7 @@ class CustomChart:
         axis_x.setRange(*axis_x_range)
         # axis_x.setLabelsVisible(False)
         axis_x.setLabelsFont(QtGui.QFont("Arial", 6))
-        axis_x.setTickCount(sample_count / 4)
+        axis_x.setTickCount(int(sample_count / 4))
 
         axis_y = self.chart.axes(QtCore.Qt.Vertical)[0]
         axis_y.setTitleText(axis_y_label)
